# internal/api/route/route.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Route public API",
    "version": "2.0",
    "contact": {
      "name": "E",
      "url": "https://blog.bullgare.com/2020/07/complete-list-of-swagger-options-to-protobuf-file/"
    }
  },
  "tags": [
    {
      "name": "RouteService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/route/add": {
      "post": {
        "summary": "AddRoute добавить роут",
        "operationId": "RouteService_AddRoute",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeAddRouteRequest"
            }
          }
        ],
        "tags": [
          "RouteService"
        ]
      }
    },
    "/v2/route/remove": {
      "post": {
        "summary": "RemoveRoute удалить роут",
        "operationId": "RouteService_RemoveRoute",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/routeRemoveRouteRequest"
            }
          }
        ],
        "tags": [
          "RouteService"
        ]
      }
    },
    "/v2/route/state": {
      "get": {
        "summary": "GetState вернуть все роуты",
        "operationId": "RouteService_GetState",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/routeGetStateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "RouteService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "routeAddRouteRequest": {
      "type": "object",
      "properties": {
        "hcDestIP": {
          "type": "string"
        },
        "hcTunDestIP": {
          "type": "string"
        }
      },
      "title": "AddRouteRequest добавить роут"
    },
    "routeGetStateResponse": {
      "type": "object",
      "properties": {
        "routes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/routeRoute"
          },
          "title": "routes список роутов"
        }
      },
      "title": "GetStateResponse выдаем все роуты"
    },
    "routeRemoveRouteRequest": {
      "type": "object",
      "properties": {
        "hcDestIP": {
          "type": "string"
        },
        "hcTunDestIP": {
          "type": "string"
        }
      },
      "title": "AddRouteRequest добавить роут"
    },
    "routeRoute": {
      "type": "object",
      "properties": {
        "hcDestIP": {
          "type": "string"
        },
        "hcTunDestIP": {
          "type": "string"
        },
        "dev": {
          "type": "string",
          "title": "dev net-link device"
        },
        "table": {
          "type": "string",
          "title": "table route table"
        }
      },
      "title": "Route IP route"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}